Support for files up to 2 GB

+++ HttpGw/src/main/java/FileMetaData.java
@@ -1,6 +1,6 @@
 public class FileMetaData {
     private final boolean exists;
-    private final long size;
+    private final int size;
     private final String type;
 
     public FileMetaData(String content) {
@@ -12,7 +12,7 @@ public class FileMetaData {
         else {
             String[] tokens = content.split(",");
             exists = Boolean.parseBoolean(tokens[0].split(":")[1]);
-            size = Long.parseLong(tokens[1].split(":")[1]);
+            size = Integer.parseInt(tokens[1].split(":")[1]);
             type = tokens[2].split(":")[1].trim();
         }
     }
@@ -21,7 +21,7 @@ public class FileMetaData {
         return exists;
     }
 
-    public long getSize() {
+    public int getSize() {
         return size;
     }
